+++ apps/expo-go/ios/Exponent/DevMenu/DevMenuManager.swift
@@ -55,7 +55,7 @@ public class DevMenuManager: NSObject {
 
   var currentBundleURL: URL? {
     guard let manifest = currentManifest else { return nil }
-    return ApiUtil.bundleUrlFromManifest(manifest)
+    return ApiUtil.bundleUrlFromManifest(manifest, relativeTo: currentManifestURL)
   }
 
   var hasActiveApp: Bool {